requirements.txt
altair==5.5.0
asgiref==3.7.2
asttokens==2.4.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.1
CMax==5.14
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==45.0.6
cvlib==0.2.7
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dj-database-url==2.3.0
Django==4.2.4
django-heroku==0.3.1
dlib @ file:///C:/Users/KHAJA/Downloads/dlib-python%20whl%20packages/dlib-python%20whl%20packages/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
docx2txt==0.9
executing==2.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
fonttools==4.53.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.7
imageio==2.34.2
imutils==0.5.4
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
language_tool_python==2.9.4
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
narwhals==2.0.1
nest-asyncio==1.6.0
nodejs==0.1.1
numpy==1.25.2
opencv-contrib-python==4.10.0.84
opencv-python==4.8.0.76
optional-django==0.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfminer.six==20250506
pillow==10.4.0
platformdirs==4.2.2
plotly==6.2.0
progressbar==2.5
prompt_toolkit==3.0.47
protobuf==6.31.1
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.26.3
pyparsing==3.1.2
pyspellchecker==0.8.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2024.1
pywin32==306
pyzmq==26.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
smmap==5.0.2
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.48.0
tenacity==9.1.2
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.2
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.8.2
